<refactor>(<edu.nextstep.racing>): MVC 패턴 반영
