+++ src/main/java/edu/nextstep/racing/domain/MovingCheck.java
@@ -1,4 +1,4 @@
-package edu.nextstep.racing;
+package edu.nextstep.racing.domain;
 
 public interface MovingCheck {
     boolean isMovable();
